feat: extracted firebase initialization to it's own lib/firebaseClient.ts file
