+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){const e=e=>{"require"!==e&&(e+=".js");let s=Promise.resolve();return i[e]||(s=new Promise((async s=>{if("document"in self){const i=document.createElement("script");i.src=e,document.head.appendChild(i),i.onload=s}else importScripts(e),s()}))),s.then((()=>{if(!i[e])throw new Error(`Module ${e} didn’t register its module`);return i[e]}))},s=(s,i)=>{Promise.all(s.map(e)).then((e=>i(1===e.length?e[0]:e)))},i={require:Promise.resolve(s)};self.define=(s,n,c)=>{i[s]||(i[s]=Promise.resolve().then((()=>{let i={};const a={uri:location.origin+s.slice(1)};return Promise.all(n.map((s=>{switch(s){case"exports":return i;case"module":return a;default:return e(s)}}))).then((e=>{const s=c(...e);return i.default||(i.default=s),i}))})))}}define("./sw.js",["./workbox-8778d57b"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/MNN7d2wi_d9zXC-IlTwEV/_buildManifest.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/MNN7d2wi_d9zXC-IlTwEV/_ssgManifest.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/29107295.a1f68933abc4c662ce97.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/322a7e8b.b2681d6404ac7e0e1cc4.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/71247caf95475e3ea7f9a0f8a30beb258b23d005.cdb38c434f3ea65aba53.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/analytics.d5d0bcf69500da96badf.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/auth.f993cbc1befb4dbeb6a7.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/commons.68093b8986013cdfc385.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/d91e9ae9.8bfd28c610966486bafd.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/database.09331c5a647ed5a0d5f7.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/f6078781a05fe1bcb0902d23dbbb2662c8d200b3.f83d01680863c178f752.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/ff239f9d.d81d3baeb35b98de4ebb.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/firestore.891d604d0d856e84170e.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/framework.76fac10824b6450d7d51.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/functions.ab049b158d5b9435deda.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/main-f5371a31d89a9604e70d.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/messaging.5f1f0fdf2127575bac34.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/pages/_app-6d740630a996ff3dbcb5.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/pages/_error-4af68a2984f73b0e6762.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/pages/index-27783d2057595dd5a66f.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/performance.8ceb77878f7cc9ff7510.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/polyfills-8a3dd1ee019d5861180f.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/remoteConfig.0bb444d96130a87fdffd.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/storage.bfa3b3a64011c296575e.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/chunks/webpack-38399d92bee13558c7aa.js",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/_next/static/css/de3c0a37dfeca0d9cfaa.css",revision:"MNN7d2wi_d9zXC-IlTwEV"},{url:"/icons/android-chrome-192x192.png",revision:"983baedb62c553a3870a4b2719936150"},{url:"/icons/android-chrome-512x512.png",revision:"88aebf5d47d9c56e37d24911679d4cc6"},{url:"/icons/apple-touch-icon.png",revision:"c406dbba03494b604b04007b2da614cf"},{url:"/icons/favicon-16x16.png",revision:"90acebf02a494885f3c8425254a24f24"},{url:"/icons/favicon-32x32.png",revision:"cae6fd709fff91995cf63ffb913402be"},{url:"/icons/favicon.ico",revision:"d800ad90ef4896ec8919f8bb79b0a2dd"},{url:"/icons/icon-check.svg",revision:"5742160ca6ea943a508cc5763e8cd312"},{url:"/icons/icon-cross.svg",revision:"fe68d57f09f867dbbd1ff592c3b9a017"},{url:"/icons/icon-moon.svg",revision:"e66b8c621b1f209e44a7c955c0e5e809"},{url:"/icons/icon-sun.svg",revision:"b3182f3b0ccf1d65cae48b64861b83a9"},{url:"/icons/mstile-150x150.png",revision:"fb296382e2a3def17bde481e04f298b2"},{url:"/images/bg-desktop-dark.jpg",revision:"02105f02a8dba33f6c8bafac6a9b0979"},{url:"/images/bg-desktop-light.jpg",revision:"988cb74c73a65433889e51b150c1a8f9"},{url:"/images/bg-mobile-dark.jpg",revision:"b0112ea5222ffd8c9d69349be56c620b"},{url:"/images/bg-mobile-light.jpg",revision:"266f20c7fd7778c0b72ac804fd769fe5"},{url:"/manifest.json",revision:"3e4104272acc50d844f6f8f03694783e"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[new e.ExpirationPlugin({maxEntries:1,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis|gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\/api\/.*$/i,new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/.*/i,new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET")}));
+if(!self.define){const e=e=>{"require"!==e&&(e+=".js");let s=Promise.resolve();return n[e]||(s=new Promise((async s=>{if("document"in self){const n=document.createElement("script");n.src=e,document.head.appendChild(n),n.onload=s}else importScripts(e),s()}))),s.then((()=>{if(!n[e])throw new Error(`Module ${e} didn’t register its module`);return n[e]}))},s=(s,n)=>{Promise.all(s.map(e)).then((e=>n(1===e.length?e[0]:e)))},n={require:Promise.resolve(s)};self.define=(s,i,c)=>{n[s]||(n[s]=Promise.resolve().then((()=>{let n={};const r={uri:location.origin+s.slice(1)};return Promise.all(i.map((s=>{switch(s){case"exports":return n;case"module":return r;default:return e(s)}}))).then((e=>{const s=c(...e);return n.default||(n.default=s),n}))})))}}define("./sw.js",["./workbox-8778d57b"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/1BKeBZTPWFsdHXpGJ4sJM/_buildManifest.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/1BKeBZTPWFsdHXpGJ4sJM/_ssgManifest.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/29107295.621d742822a168cc3713.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/322a7e8b.2ae72b1a8cfd46c82f89.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/analytics.9587c9f007f778e0f536.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/auth.ce42de8c6b055515c3e0.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/commons.b7e3d74f3ae569805384.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/d91e9ae9.3bb3336b769ad9cd8937.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/database.c3a27275c72dc0ab6312.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/f6078781a05fe1bcb0902d23dbbb2662c8d200b3.2fb93d4585311b8c5370.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/ff239f9d.60ff2b3d4b36fb9866da.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/framework.76fac10824b6450d7d51.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/functions.eff3e5ba7753ca5abecd.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/main-a8063d8820f2f1c8ceac.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/messaging.f5997fef7012311db5ff.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/pages/_app-52d9cc5f3d0eab3e11a5.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/pages/_error-061ec44534b1b60fe788.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/pages/index-94a060202f69341b9865.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/performance.ba43e2c8f61e7f39fb5c.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/polyfills-4e81d2a8a9f16b342b90.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/remoteConfig.20086258e810404bb794.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/storage.5fd22ee9fb0dd50eb127.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/chunks/webpack-94e8cd221590a6eedadf.js",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/_next/static/css/de3c0a37dfeca0d9cfaa.css",revision:"1BKeBZTPWFsdHXpGJ4sJM"},{url:"/icons/android-chrome-192x192.png",revision:"983baedb62c553a3870a4b2719936150"},{url:"/icons/android-chrome-512x512.png",revision:"88aebf5d47d9c56e37d24911679d4cc6"},{url:"/icons/apple-touch-icon.png",revision:"c406dbba03494b604b04007b2da614cf"},{url:"/icons/favicon-16x16.png",revision:"90acebf02a494885f3c8425254a24f24"},{url:"/icons/favicon-32x32.png",revision:"cae6fd709fff91995cf63ffb913402be"},{url:"/icons/favicon.ico",revision:"d800ad90ef4896ec8919f8bb79b0a2dd"},{url:"/icons/icon-check.svg",revision:"5742160ca6ea943a508cc5763e8cd312"},{url:"/icons/icon-cross.svg",revision:"fe68d57f09f867dbbd1ff592c3b9a017"},{url:"/icons/icon-moon.svg",revision:"e66b8c621b1f209e44a7c955c0e5e809"},{url:"/icons/icon-sun.svg",revision:"b3182f3b0ccf1d65cae48b64861b83a9"},{url:"/icons/mstile-150x150.png",revision:"fb296382e2a3def17bde481e04f298b2"},{url:"/images/bg-desktop-dark.jpg",revision:"02105f02a8dba33f6c8bafac6a9b0979"},{url:"/images/bg-desktop-light.jpg",revision:"988cb74c73a65433889e51b150c1a8f9"},{url:"/images/bg-mobile-dark.jpg",revision:"b0112ea5222ffd8c9d69349be56c620b"},{url:"/images/bg-mobile-light.jpg",revision:"266f20c7fd7778c0b72ac804fd769fe5"},{url:"/manifest.json",revision:"3e4104272acc50d844f6f8f03694783e"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[new e.ExpirationPlugin({maxEntries:1,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis|gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\/api\/.*$/i,new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/.*/i,new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET")}));
